refactor: rename internal modules for clarity
loader → ontology_loader, compiler → graph_ddl_compiler,
resolved_models → graph_ddl_models, test_compiler → test_graph_ddl_compiler.
Public API unchanged.
